upload code thesis13.ppt, reconfigurable paxos

+++ code/heartbeat_lib.h
@@ -7,13 +7,14 @@ class HeartbeatClient{
 private:
 	char ip[IP_LENGTH];
 	int sfd;
-	unsigned timestamp;
+	double timestamp;
 public:
 	HeartbeatClient(const char *ip);
 	~HeartbeatClient();
 	int readMessage(); // return number of bytes read, 0 or -1 if tcp shutdown
-	int isExpired(unsigned t);
+	int isExpired(double t);
 	int getFD();
 	std::string getIP();
 	void getIP(char *s);
 };
+
